Fix backend bug and refactor DeleteForm
